Calculates adjusted mlm effect size

Estimates a standardized mean difference effect size from a fitted multi-level model, using adjusted mlm method as described in Pustejovsky, Hedges, & Shadish (2014).

g_mlm(mod, p_const, r_const, infotype = "expected", returnModel = TRUE)

Arguments

mod
Fitted model of class lmeStruct (estimated using nlme::lme()) or of class glsStruct (estimated using nlme::gls()).

p_const
Vector of constants for calculating numerator of effect size. Must be the same length as fixed effects in mod.

r_const
Vector of constants for calculating denominator of effect size. Must be the same length as the number of variance component parameters in mod.

infotype
Type of information matrix. One of "expected" (the default), "observed", or "average".

returnModel
(Optional) If true, the fitted input model is included in the return. Defaults to TRUE so that summary() method returns more detail about the model parameters for an object of class g_mlm.

Value

A list with the following components

p_beta	Numerator of effect size

r_theta	Squared denominator of effect size

delta_AB	Unadjusted (mlm) effect size estimate

nu	Estimated denominator degrees of freedom

J_nu	Biased correction factor for effect size estimate

kappa	Scaled standard error of numerator

g_AB	Corrected effect size estimate

SE_g_AB	Approximate standard error estimate

cnvg_warn	Indicator that model did not converge

theta	Estimated variance component parameters

info_inv	Inversed information matrix

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi: 10.3102/1076998614547577

Examples

library(nlme)
data(Bryant2016, package = "lmeInfo")
Bryant2016_RML1 <- lme(fixed = outcome ~ treatment,
                       random = ~ 1 | school/case,
                       correlation = corAR1(0, ~ session | school/case),
                       data = Bryant2016)
Bryant2016_g1 <- g_mlm(Bryant2016_RML1, p_const = c(0,1), r_const = c(1,1,0,1),
                       infotype = "expected", returnModel = TRUE)
summary(Bryant2016_g1)

#>                                            est      se
#> Tau.school.school.var((Intercept))     105.619 285.200
#> Tau.case.case.var((Intercept))           0.000 632.159
#> cor_params                               0.978   0.015
#> sigma_sq                              1058.303 678.388
#> total variance                        1163.922 366.515
#> (Intercept)                             84.571   9.684
#> treatmenttreatment                      16.425   2.975
#> treatment effect at a specified time    16.425   2.975
#> unadjusted effect size                   0.481   0.122
#> adjusted effect size                     0.463   0.118
#> degree of freedom                       20.169        
#> constant kappa                           0.087        
#> logLik                               -1133.122        

print(Bryant2016_g1)

#>                           est    se
#> unadjusted effect size  0.481 0.122
#> adjusted effect size    0.463 0.118
#> degree of freedom      20.169      

data(Laski, package = "scdhlm")
Laski_AR1 <- gls(outcome ~ treatment,
                 correlation = corAR1(0.2, ~ time | case),
                 data = Laski)
Laski_AR1_g <- g_mlm(Laski_AR1, p_const = c(0,1), r_const = c(0,1),
                     infotype = "expected", returnModel = TRUE)
summary(Laski_AR1_g)

#>                                           est     se
#> cor_params                              0.646  0.070
#> sigma_sq                              414.141 80.575
#> total variance                        414.141 80.575
#> (Intercept)                            40.175  4.363
#> treatmenttreatment                     29.723  4.894
#> treatment effect at a specified time   29.723  4.894
#> unadjusted effect size                  1.461  0.285
#> adjusted effect size                    1.440  0.281
#> degree of freedom                      52.836       
#> constant kappa                          0.240       
#> logLik                               -529.156       

print(Laski_AR1_g)

#>                           est    se
#> unadjusted effect size  1.461 0.285
#> adjusted effect size    1.440 0.281
#> degree of freedom      52.836