Estimated sampling variance-covariance of variance component parameters.

Estimate the sampling variance-covariance of variance component parameters from a fitted linear mixed effects model (lmeStruct object) or generalized least squares model (glsStruct object) using the inverse Fisher information.

varcomp_vcov(mod, type = "expected")

Arguments

mod
Fitted model of class lmeStruct or glsStruct.

type
Type of information matrix. One of "expected" (the default), "observed", or "average".

Value

Sampling variance-covariance matrix corresponding to variance component parameters of mod.

Examples

library(nlme)
data(Bryant2016)
Bryant2016_RML <- lme(fixed = outcome ~ treatment,
                      random = ~ 1 | school/case,
                      correlation = corAR1(0, ~ session | school/case),
                      data = Bryant2016)
varcomp_vcov(Bryant2016_RML, type = "expected")

#>                             Tau.school.var((Intercept))
#> Tau.school.var((Intercept))                8.133900e+04
#> Tau.case.var((Intercept))                 -2.140289e+04
#> cor_params                                -1.004342e-02
#> sigma_sq                                  -4.701744e+02
#>                             Tau.case.var((Intercept))    cor_params
#> Tau.school.var((Intercept))             -2.140289e+04 -0.0100434219
#> Tau.case.var((Intercept))                3.996251e+05 -8.2280433012
#> cor_params                              -8.228043e+00  0.0002154736
#> sigma_sq                                -3.815479e+05  9.8729751627
#>                                  sigma_sq
#> Tau.school.var((Intercept)) -4.701744e+02
#> Tau.case.var((Intercept))   -3.815479e+05
#> cor_params                   9.872975e+00
#> sigma_sq                     4.602107e+05